(feat:vectors): calc, add token in db
